refactor(recruitment): separate the package
